Extract out the duo upgrades and croissant generated into an independent script that can be ran assuming you pipe in all of the independent data fields.
